Extract `VimModeSetting` to its own crate
