+++ crates/vim_mode_setting/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "vim_mode_setting"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/vim_mode_setting.rs"
+
+[dependencies]
+anyhow.workspace = true
+gpui.workspace = true
+settings.workspace = true
