+++ crates/vim_mode_setting/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
