+++ crates/welcome/Cargo.toml
@@ -30,7 +30,7 @@ settings.workspace = true
 theme_selector.workspace = true
 ui.workspace = true
 util.workspace = true
-vim.workspace = true
+vim_mode_setting.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
