+++ crates/zed/Cargo.toml
@@ -119,6 +119,7 @@ util.workspace = true
 uuid.workspace = true
 vcs_menu.workspace = true
 vim.workspace = true
+vim_mode_setting.workspace = true
 welcome.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
